refactoring and reshaping of dtos

+++ HandiPapi/Repository/GenericRepository.cs
@@ -7,18 +7,19 @@ namespace HandiPapi.Repository
 {
     public class GenericRepository<T> : IGenericRepository<T> where T : class
     {
-        private readonly DatabaseContext context;
+        private readonly DatabaseContext _context;
         private readonly DbSet<T> db;
 
         public GenericRepository(DatabaseContext context)
         {
-            this.context = context;
-            this.db = context.Set<T>();
+            _context = context;
+            db = context.Set<T>();
         }
         public async Task Delete(int Id)
         {
             var entity = await db.FindAsync(Id);
-            db.Remove(entity);
+            if (entity != null)
+                _ = db.Remove(entity);
         }
 
         public void DeleteRange(IEnumerable<T> entities)
@@ -26,7 +27,7 @@ namespace HandiPapi.Repository
             db.RemoveRange(entities);
         }
 
-        public async Task<T> Get(Expression<Func<T, bool>> expression, List<string> include = null)
+        public async Task<T?> Get(Expression<Func<T, bool>> expression, List<string>? include = null)
         {
             IQueryable<T> query = db;
             if(include != null)
@@ -39,15 +40,14 @@ namespace HandiPapi.Repository
             return await query.AsNoTracking().FirstOrDefaultAsync(expression);
         }
 
-        public async Task<IList<T>> GetAll(Expression<Func<T, bool>> expression = null, Func<IQueryable<T>, IOrderedQueryable<T>> ordeyBy = null, List<string> include = null)
+        public async Task<IList<T>> GetAll(Expression<Func<T, bool>>? expression = null, Func<IQueryable<T>, IOrderedQueryable<T>>? ordeyBy = null, List<string>? include = null)
         {
             IQueryable<T> query = db;
             if(expression != null)
             {
-                foreach(var includeValue in include)
-                {
-                    query = query.Where(expression);
-                }
+
+                query = query.Where(expression);
+                
             }
 
             if(include != null)
@@ -78,7 +78,7 @@ namespace HandiPapi.Repository
         public void Update(T entity)
         {
             db.Attach(entity);
-            context.Entry(entity).State = EntityState.Modified;
+            _context.Entry(entity).State = EntityState.Modified;
         }
     }
 }
